requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-decouple>=3.8
django-tailwind>=3.8
Pillow>=10.0
gunicorn>=21.0
whitenoise>=6.6
django-extensions>=3.2
django-debug-toolbar>=4.2
